+++ src/internal/streamstore/collectables_test.go
@@ -64,6 +64,11 @@ func (suite *StreamStoreIntgSuite) TearDownSubTest() {
 }
 
 func (suite *StreamStoreIntgSuite) TestStreamer() {
+	deetsPath, err := path.FromDataLayerPath("tenant-id/exchange/user-id/email/Inbox/folder1/foo", true)
+	require.NoError(suite.T(), err, clues.ToCore(err))
+
+	locPath := path.Builder{}.Append(deetsPath.Folders()...)
+
 	table := []struct {
 		name      string
 		deets     func(*testing.T) *details.Details
@@ -81,12 +86,15 @@ func (suite *StreamStoreIntgSuite) TestStreamer() {
 			deets: func(t *testing.T) *details.Details {
 				deetsBuilder := &details.Builder{}
 				require.NoError(t, deetsBuilder.Add(
-					"rr", "sr", "pr", "lr",
+					deetsPath,
+					locPath,
 					true,
 					details.ItemInfo{
-						Exchange: &details.ExchangeInfo{Subject: "hello world"},
+						Exchange: &details.ExchangeInfo{
+							ItemType: details.ExchangeMail,
+							Subject:  "hello world",
+						},
 					}))
-
 				return deetsBuilder.Details()
 			},
 			errs:      func() *fault.Errors { return nil },
@@ -112,10 +120,14 @@ func (suite *StreamStoreIntgSuite) TestStreamer() {
 			deets: func(t *testing.T) *details.Details {
 				deetsBuilder := &details.Builder{}
 				require.NoError(t, deetsBuilder.Add(
-					"rr", "sr", "pr", "lr",
+					deetsPath,
+					locPath,
 					true,
 					details.ItemInfo{
-						Exchange: &details.ExchangeInfo{Subject: "hello world"},
+						Exchange: &details.ExchangeInfo{
+							ItemType: details.ExchangeMail,
+							Subject:  "hello world",
+						},
 					}))
 
 				return deetsBuilder.Details()
@@ -179,6 +191,7 @@ func (suite *StreamStoreIntgSuite) TestStreamer() {
 				assert.Equal(t, deets.Entries[0].ShortRef, readDeets.Entries[0].ShortRef)
 				assert.Equal(t, deets.Entries[0].RepoRef, readDeets.Entries[0].RepoRef)
 				assert.Equal(t, deets.Entries[0].LocationRef, readDeets.Entries[0].LocationRef)
+				assert.Equal(t, deets.Entries[0].ItemRef, readDeets.Entries[0].ItemRef)
 				assert.Equal(t, deets.Entries[0].Updated, readDeets.Entries[0].Updated)
 				assert.NotNil(t, readDeets.Entries[0].Exchange)
 				assert.Equal(t, *deets.Entries[0].Exchange, *readDeets.Entries[0].Exchange)