+++ src/pkg/selectors/exchange.go
@@ -2,6 +2,7 @@ package selectors
 
 import (
 	"context"
+	"fmt"
 	"strconv"
 
 	"github.com/alcionai/clues"
@@ -120,6 +121,15 @@ func (s exchange) PathCategories() selectorPathCategories {
 		Includes: pathCategoriesIn[ExchangeScope, exchangeCategory](s.Includes),
 	}
 }
+
+// ---------------------------------------------------------------------------
+// Stringers and Concealers
+// ---------------------------------------------------------------------------
+
+func (s ExchangeScope) Conceal() string             { return conceal(s) }
+func (s ExchangeScope) Format(fs fmt.State, r rune) { format(s, fs, r) }
+func (s ExchangeScope) String() string              { return conceal(s) }
+func (s ExchangeScope) PlainString() string         { return plainString(s) }
 
 // -------------------
 // Exclude/Includes
@@ -336,7 +346,7 @@ func (sr *ExchangeRestore) ContactName(senderID string) []ExchangeScope {
 			ExchangeContact,
 			ExchangeInfoContactName,
 			[]string{senderID},
-			wrapSliceFilter(filters.In)),
+			filters.In),
 	}
 }
 
@@ -351,7 +361,7 @@ func (sr *ExchangeRestore) EventOrganizer(organizer string) []ExchangeScope {
 			ExchangeEvent,
 			ExchangeInfoEventOrganizer,
 			[]string{organizer},
-			wrapSliceFilter(filters.In)),
+			filters.In),
 	}
 }
 
@@ -366,7 +376,7 @@ func (sr *ExchangeRestore) EventRecurs(recurs string) []ExchangeScope {
 			ExchangeEvent,
 			ExchangeInfoEventRecurs,
 			[]string{recurs},
-			wrapFilter(filters.Equal)),
+			filters.Equal),
 	}
 }
 
@@ -380,7 +390,7 @@ func (sr *ExchangeRestore) EventStartsAfter(timeStrings string) []ExchangeScope 
 			ExchangeEvent,
 			ExchangeInfoEventStartsAfter,
 			[]string{timeStrings},
-			wrapFilter(filters.Less)),
+			filters.Less),
 	}
 }
 
@@ -394,7 +404,7 @@ func (sr *ExchangeRestore) EventStartsBefore(timeStrings string) []ExchangeScope
 			ExchangeEvent,
 			ExchangeInfoEventStartsBefore,
 			[]string{timeStrings},
-			wrapFilter(filters.Greater)),
+			filters.Greater),
 	}
 }
 
@@ -409,7 +419,7 @@ func (sr *ExchangeRestore) EventSubject(subject string) []ExchangeScope {
 			ExchangeEvent,
 			ExchangeInfoEventSubject,
 			[]string{subject},
-			wrapSliceFilter(filters.In)),
+			filters.In),
 	}
 }
 
@@ -423,7 +433,7 @@ func (sr *ExchangeRestore) MailReceivedAfter(timeStrings string) []ExchangeScope
 			ExchangeMail,
 			ExchangeInfoMailReceivedAfter,
 			[]string{timeStrings},
-			wrapFilter(filters.Less)),
+			filters.Less),
 	}
 }
 
@@ -437,7 +447,7 @@ func (sr *ExchangeRestore) MailReceivedBefore(timeStrings string) []ExchangeScop
 			ExchangeMail,
 			ExchangeInfoMailReceivedBefore,
 			[]string{timeStrings},
-			wrapFilter(filters.Greater)),
+			filters.Greater),
 	}
 }
 
@@ -452,7 +462,7 @@ func (sr *ExchangeRestore) MailSender(sender string) []ExchangeScope {
 			ExchangeMail,
 			ExchangeInfoMailSender,
 			[]string{sender},
-			wrapSliceFilter(filters.In)),
+			filters.In),
 	}
 }
 
@@ -467,7 +477,7 @@ func (sr *ExchangeRestore) MailSubject(subject string) []ExchangeScope {
 			ExchangeMail,
 			ExchangeInfoMailSubject,
 			[]string{subject},
-			wrapSliceFilter(filters.In)),
+			filters.In),
 	}
 }
 
@@ -584,6 +594,7 @@ func (ec exchangeCategory) isLeaf() bool {
 func (ec exchangeCategory) pathValues(
 	repo path.Path,
 	ent details.DetailsEntry,
+	cfg Config,
 ) (map[categorizer][]string, error) {
 	var folderCat, itemCat categorizer
 
@@ -601,9 +612,14 @@ func (ec exchangeCategory) pathValues(
 		return nil, clues.New("bad exchanageCategory").With("category", ec)
 	}
 
+	item := ent.ItemRef
+	if len(item) == 0 {
+		item = repo.Item()
+	}
+
 	result := map[categorizer][]string{
 		folderCat: {repo.Folder(false)},
-		itemCat:   {repo.Item(), ent.ShortRef},
+		itemCat:   {item, ent.ShortRef},
 	}
 
 	if len(ent.LocationRef) > 0 {